pyproject.toml
[project]
name = "career-caddy-ai"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "pydantic-ai>=0.0.14",
    "playwright>=1.40.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "fastmcp>=3.0.0,<4",
    "rich<14",
    "html2text",
    "beautifulsoup4",
    # logfire[httpx] pulls opentelemetry-instrumentation-httpx so
    # logfire.instrument_httpx() actually wires in. Without the extra
    # the instrumenter fails at runtime and tools silently skip HTTP
    # spans.
    "logfire[httpx]",
    "pyyaml",
    "camoufox>=0.4.11",
    "playwright-stealth>=1.0.6",
    "jinja2>=3.0.0",
    "starlette>=0.37.0",
]

[project.scripts]
caddy-pipeline = "agents.job_email_to_caddy:run"
caddy-public   = "mcp_servers.public_server:main"
caddy-chat     = "mcp_servers.chat_server:main"
caddy-poller  = "pollers.hold_poller:main_sync"
caddy-score    = "pollers.score_poller:run"
caddy-fetch-browser = "camoufox.__main__:main"
caddy-fetch-chromium = "tools.fetch_chromium:main"
caddy-export-graph = "tools.export_graph_structure:main"

[tool.hatch.build.targets.wheel]
packages = ["agents", "mcp_servers", "browser", "scrape_graph", "pollers", "tools", "lib"]

[tool.hatch.build.targets.editable]
path = "."

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]
